WNB-181: Start moving store folder to TS
