+++ package.json
@@ -39,6 +39,15 @@
     "@babel/preset-env": "7.4.3",
     "@babel/preset-react": "7.0.0",
     "@babel/preset-stage-3": "7.0.0",
+    "@types/jest": "24.0.23",
+    "@types/node": "12.12.14",
+    "@types/ramda": "0.26.36",
+    "@types/react": "16.9.13",
+    "@types/react-dom": "16.9.4",
+    "@types/react-redux": "7.1.5",
+    "@types/react-router-dom": "5.1.3",
+    "@types/redux-actions": "2.6.1",
+    "@types/styled-components": "4.4.0",
     "awesome-typescript-loader": "5.2.1",
     "babel-cli": "6.26.0",
     "babel-eslint": "10.0.1",
@@ -56,7 +65,7 @@
     "eslint-plugin-babel": "5.3.0",
     "eslint-plugin-import": "2.14.0",
     "eslint-plugin-jsx-a11y": "6.1.1",
-    "eslint-plugin-prettier": "3.0.1",
+    "eslint-plugin-prettier": "3.1.1",
     "eslint-plugin-react": "7.11.0",
     "eslint-plugin-react-hooks": "2.1.2",
     "file-loader": "3.0.1",
@@ -66,7 +75,7 @@
     "lint-staged": "8.1.5",
     "mini-css-extract-plugin": "0.5.0",
     "node-sass": "4.12.0",
-    "prettier": "1.16.4",
+    "prettier": "1.19.1",
     "react-hot-loader": "4.8.3",
     "redux": "4.0.1",
     "sass-loader": "7.1.0",
@@ -80,14 +89,6 @@
     "webpack-md5-hash": "0.0.6"
   },
   "dependencies": {
-    "@types/jest": "24.0.23",
-    "@types/node": "12.12.14",
-    "@types/ramda": "0.26.36",
-    "@types/react": "16.9.13",
-    "@types/react-dom": "16.9.4",
-    "@types/react-redux": "7.1.5",
-    "@types/react-router-dom": "5.1.3",
-    "@types/styled-components": "4.4.0",
     "axios": "0.19.0",
     "history": "4.10.1",
     "prop-types": "15.7.2",
