+++ src/store/auth/changePassword/types.ts
@@ -1,0 +1,3 @@
+import { EntityInitialState } from '../../../types';
+
+export interface ChangePasswordState extends EntityInitialState {}
