+++ src/store/auth/forgotPassword/actionCreators.ts
@@ -5,10 +5,11 @@ import {
   FORGOT_PASSWORD_SUCCESS,
   FORGOT_PASSWORD_FAILURE,
 } from './constants';
+import { ForgotPasswordRequestPayload } from './types';
 
 export const forgotPassword = createAction(
   FORGOT_PASSWORD_REQUEST,
-  payload => payload
+  (payload: ForgotPasswordRequestPayload) => payload
 );
 export const forgotPasswordSuccess = createAction(FORGOT_PASSWORD_SUCCESS);
 export const forgotPasswordFailure = createAction(FORGOT_PASSWORD_FAILURE);
