+++ src/store/auth/forgotPassword/types.ts
@@ -1,0 +1,7 @@
+import { EntityInitialState } from '../../../types';
+
+export interface ForgotPasswordState extends EntityInitialState {}
+
+export interface ForgotPasswordRequestPayload {
+  [key: string]: any;
+}
