+++ src/types.d.ts
@@ -1,0 +1,12 @@
+export interface EntityInitialState {
+  loading: boolean;
+  loaded: boolean;
+  error: string;
+}
+
+export interface Profile {
+  firstName: string;
+  lastName: string;
+  phone: string;
+  picture: string;
+}
